refactor: Update request models for collaborators & projects
- Replace AddUpdateProjectRequest with CreateUpdateProjectRequest that validates project names.
- Move CreateUpdateProjectRequest to Requests.Projects namespace.
- Move AddCollaboratorRequest to Requests.Collaborators namespace.
- Change dedicated namespace for requests in CollaboratorsController and ProjectsController.
